feat: 로그인된 상태에서 인증페이지로 이동하면 홈으로 리다이렉트되는 로직 추가 JUL-89
